+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Fix serialization of `ThreeDSecureLookupAdditionalInformation`
 * Return `AuthenticationInsight` when requested on `PaymentMethodNonce.create`
 
 ## 2.96.0